+++ src/dto/thrid/hook.go
@@ -3,6 +3,6 @@ package thrid
 // HookSendReq ...
 type HookSendReq struct {
 	Event   string   `json:"event"`
-	Users   []string `json:"users"`
+	Users   []string `json:"users,omitempty"`
 	Content string   `json:"content"`
 }